services/doc-ingest-service/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==1.10.18
httpx==0.27.2
SQLAlchemy==2.0.35
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
pytest==8.3.3
